chore: nodemailer refactor working
