+++ services/auth/confirm.js
@@ -1,5 +1,5 @@
 const { confirmUser } = require('../../query/auth')
-const { mailer } = require('../../config/mailer')
+const { sendMail } = require('../../config/mailer')
 
 module.exports = (db) => async (req, res, next) => {
   const { token } = req.params
@@ -14,9 +14,7 @@ module.exports = (db) => async (req, res, next) => {
   }
 
   // sending confirmation email
-  await mailer(
-    { to: result.email, username: result.username }
-  )
+  await sendMail.confirmation({ to: result.email, username: result.username })
 
   res.status(200)
     .json({
