+++ utils/index.js
@@ -1,5 +1,16 @@
 const upperCaseFn = str => str.replace(/(^\w|\s\w)/g, m => m.toUpperCase())
 
+const catcher = (fn) => async ({...args}) => {
+    try{
+      await fn({...args})
+    } catch (e) {
+      console.info('Error at query: ', error.message)
+      return false
+    }
+  }
+  
+
 module.exports = {
-    upperCaseFn
+    upperCaseFn,
+    catcher,
 }
